% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ccw.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate weights for each arm of a long data.frame}
\usage{
calc_weights(df, event_times_df, predvars)
}
\arguments{
\item{df}{the data.frame for a single arm}

\item{event_times_df}{the event times data.frame}

\item{predvars}{the baseline variables for adjustment}
}
\value{
a data.frame with weight columns included
}
\description{
Calculate weights for each arm of a long data.frame
}
