% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clones.R
\name{create_clones}
\alias{create_clones}
\title{Create clones for CCW analysis}
\usage{
create_clones(
  df,
  id,
  event,
  time_to_event,
  exposure,
  time_to_exposure,
  ced_window = NULL
)
}
\arguments{
\item{df}{A data.frame with one row per patient.}

\item{id}{The name of the column in \code{df} that contains the patient identifier.}

\item{event}{The name of the column in \code{df} that contains the event of interest.}

\item{time_to_event}{The name of the column in \code{df} that contains the time to event.}

\item{exposure}{The name of the column in \code{df} that contains the exposure.}

\item{time_to_exposure}{The name of the column in \code{df} that contains the time to exposure.}

\item{ced_window}{The date at which the clinical eligibility window closes. Can be left empty, in which case the clinical eligibility window is assumed to be part of
\code{exposure} and \code{time_to_exposure}}
}
\value{
A data.frame with one row per clone.
}
\description{
Pass a one-row-per-patient data.frame and get back a data.frame with one row per clone.
}
\examples{

# Load the toy dataset
data(toy_df)

# Create clones
clones <- create_clones(toy_df, 
                        id = "id", 
                        event = "death", 
                        time_to_event = "fup_obs", 
                        exposure = "surgery", 
                        time_to_exposure = "timetosurgery", 
                        ced_window = 365.25/2)
}
\references{
Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.
}
