% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_final}
\alias{data_final}
\title{Testing data for \code{cast_clones_to_long}}
\format{
An object of class \code{data.frame} with 22772 rows and 23 columns.

An object of class \code{data.frame} with 22772 rows and 23 columns.
}
\source{
\url{https://doi.org/10.1093/ije/dyaa057}

\url{https://doi.org/10.1093/ije/dyaa057}
}
\usage{
data_final

data_final
}
\description{
The long clones dataset from Maringe et al. (2020) named \code{data_final}

The long clones dataset from Maringe et al. (2020) named \code{data_final}
}
\references{
Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.

Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.
}
\keyword{datasets}
