% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_long_2}
\alias{data_long_2}
\title{Testing data for \code{calc_weights}, no surgery recipients}
\format{
An object of class \code{data.frame} with 8792 rows and 27 columns.
}
\source{
\url{https://doi.org/10.1093/ije/dyaa057}
}
\usage{
data_long_2
}
\description{
The long clones dataset with weights for no surgery recipients from Maringe et al. (2020) named \code{data_long_2}
}
\references{
Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.
}
\keyword{datasets}
