% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maringe_tab}
\alias{maringe_tab}
\title{Toy dataset from Maringe et al. (2020)}
\format{
\subsection{\code{maringe_tab}}{

A data frame with 200 rows and 12 columns:
\describe{
\item{id}{patient identifier}
\item{fup_obs}{observed follow-up time (time to death or 1 year if censored alive)}
\item{death}{observed event of interest (all-cause death) 1: dead, 0:alive}
...
}
}
}
\source{
\url{https://doi.org/10.1093/ije/dyaa057}
}
\usage{
maringe_tab
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\references{
Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.
}
\keyword{datasets}
