% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccw_over_time.R
\name{plot_ccw_over_time}
\alias{plot_ccw_over_time}
\title{Visualize weights and variables over time}
\usage{
plot_ccw_over_time(df)
}
\arguments{
\item{df}{A data.frame with one row per clone per observation period that contains weights for each patient-period as returned by \code{\link[=generate_ccw]{generate_ccw()}}}
}
\value{
a ggplot2 object
}
\description{
Visualize weights and variables over time
}
\examples{

# Load the toy dataset
data(toy_df)

# Create clones
clones <- create_clones(toy_df, 
                        id = "id", 
                        event = "death", 
                        time_to_event = "fup_obs", 
                        exposure = "surgery", 
                        time_to_exposure = "timetosurgery", 
                        ced_window = 365.25/2)

clones_long <- cast_clones_to_long(clones)
clones_long_w <- generate_ccw(clones_long, predvars = c("age"))
plot_ccw_over_time(clones_long_w)
}
