% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_mean_over_time.R
\name{plot_var_mean_over_time}
\alias{plot_var_mean_over_time}
\title{Plot a variable over time for each clone, weighted and unweighted}
\usage{
plot_var_mean_over_time(df, var)
}
\arguments{
\item{df}{A data.frame with one row per clone per observation period that contains weights for each patient-period as returned by \code{\link[=generate_ccw]{generate_ccw()}}}

\item{var}{A variable in colnames(df) to plot over time}
}
\value{
a ggplot2 object
}
\description{
Plot a variable over time for each clone, weighted and unweighted
}
\examples{

# Load the toy dataset
data(toy_df)

# Create clones
clones <- create_clones(toy_df, 
                        id = "id", 
                        event = "death", 
                        time_to_event = "fup_obs", 
                        exposure = "surgery", 
                        time_to_exposure = "timetosurgery", 
                        ced_window = 365.25/2)

clones_long <- cast_clones_to_long(clones)
clones_long_w <- generate_ccw(clones_long, predvars = c("age"))
plot_var_mean_over_time(clones_long_w, "sex")
}
