% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab}
\alias{tab}
\title{Testing data for \code{create_clones}}
\format{
An object of class \code{data.frame} with 400 rows and 6 columns.
}
\source{
\url{https://doi.org/10.1093/ije/dyaa057}
}
\usage{
tab
}
\description{
The clones dataset from Maringe et al. (2020) named \code{tab}
}
\references{
Maringe, Camille, et al. "Reflection on modern methods: trial emulation in the presence of immortal-time bias. Assessing the benefit of major surgery for elderly lung cancer patients using observational data." International journal of epidemiology 49.5 (2020): 1719-1729.
}
\keyword{datasets}
