% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ccw_calc_weights.R
\name{generate_ccw_calc_weights}
\alias{generate_ccw_calc_weights}
\title{Calculate weights for each arm of a long data.frame}
\usage{
generate_ccw_calc_weights(df, model_fmla, event_times_df, predvars)
}
\arguments{
\item{df}{the data.frame for a single arm}

\item{model_fmla}{the formula for the Cox model}

\item{event_times_df}{the event times data.frame}

\item{predvars}{the baseline variables for adjustment}
}
\value{
a data.frame with weight columns included
}
\description{
Calculate weights for each arm of a long data.frame
}
